get snap release number from tito rel-eng
Instead of prefixing the snap release with "-0", try to read the
release number from rel-eng.

Without this patch a tito package specification like this:
1.0.1-1 ./
results in a snap version with like this:
1.0.1-0.00.snap.20140615.131041.git.305fd98.git.0.15c9f3a

Which results in the snap version always being older than the
current tagged version and to upgrade to the latest snap you
have to "yum downgrade" to it.

Similarly if you have a snap version like this installed and try
to upgrade to a newer snap yum will update to the latest tagged version,
which is effectively a downgrade.